splitted models in seperate file, reformatting and permissions are still a mess
