+++ PrincessPaperplane/core/models/banned_channel.py
@@ -1,0 +1,9 @@
+from pony.orm import PrimaryKey, Required
+
+from core import database
+
+
+class BannedChannel(database.Entity):
+    _table_ = "level_banned_channel"
+    id = PrimaryKey(int, auto=True)
+    channel = Required(str)
