+++ PrincessPaperplane/core/models/level_reward.py
@@ -1,0 +1,10 @@
+from pony.orm import PrimaryKey, Required
+
+from core import database
+
+
+class LevelReward(database.Entity):
+    _table_ = "level_reward"
+    id = PrimaryKey(int, auto=True)
+    reward_level = Required(int)
+    reward_role = Required(str)
