+++ PrincessPaperplane/utility/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["Checks", "Terminal"]
+
+from utility.terminal import Terminal
+from utility.checks import Checks
