+++ PrincessPaperplane/utility/checks.py
@@ -1,9 +1,20 @@
 from typing import Optional
 
 from discord.ext import commands
+from discord.ext.commands import Context
+
+from core.models.permission import Permission
 
 
 class Checks:
+
+    @staticmethod
+    def has_permission_for(permission: str):
+        async def predicate(ctx: Context):
+            return permission in Permission.get_permissions_for_user(ctx.author)
+
+        return commands.check(predicate)
+
     @staticmethod
     def is_not_in_list(user_list: list):
         async def predicate(ctx):
