+++ README.md
@@ -1,4 +1,5 @@
 # DiscordBots
+
 [![CodeQL](https://github.com/Schrotty/DiscordBots/actions/workflows/codeql-analysis.yml/badge.svg)](https://github.com/Schrotty/DiscordBots/actions/workflows/codeql-analysis.yml)
 
 These are Discord-Bots created in Python 3 to provide features like Voicechannel-creation and personalized bots like the
